# docs/swagger.yaml
basePath: /api
definitions:
...
  model.Element:
    properties:
      content:
        description: 消息内容（文本内容或文件 HASH）
        type: string
      hash:
        description: 文件信息（仅在 Type 为图片/视频/文件时有效）
        type: string
      height:
        type: integer
      name:
        type: string
      size:
        type: integer
      type:
        allOf:
        - $ref: '#/definitions/model.MessageType'
        description: 消息类型（0：文本，1：图片，2：视频，3：文件）
      url:
        type: string
      width:
        type: integer
    type: object
...
  model.MessageType:
    enum:
    - 0
    - 1
    - 2
    - 3
    type: integer
    x-enum-varnames:
    - MsgTypeText
    - MsgTypeImage
    - MsgTypeVideo
    - MsgTypeFile
  model.Moment:
    properties:
      createdAt:
        description: 创建时间
        type: string
      elements:
        description: 朋友圈元素列表（JSON 序列化存储）
        items:
          $ref: '#/definitions/model.Element'
        type: array
      id:
        description: 朋友圈 ID
        type: integer
      likeCount:
        description: 点赞数
        type: integer
      ownerId:
        description: 朋友圈发布者 ID
        type: integer
      status:
        description: 朋友圈状态，0-正常，1-删除
        type: integer
      updatedAt:
        description: 更新时间
        type: string
      visible:
        allOf:
        - $ref: '#/definitions/model.MomentVisbileStatus'
        description: 可见性，0-公开，1-好友可见，2-仅自己可见 3-部分好友可见
    type: object
  model.MomentComments:
    properties:
      createdAt:
        description: 创建时间
        type: string
      id:
        description: 评论ID
        type: integer
      momentId:
        description: 朋友圈记录ID
        type: integer
      status:
        description: 评论状态，0-正常，1-删除
        type: integer
      updatedAt:
        description: 更新时间
        type: string
      userId:
        description: 评论者用户ID
        type: integer
      visible:
        description: 评论内容
        type: string
    type: object
  model.MomentCommentsResponse:
    properties:
      avatar:
        description: 用户头像、创建时；为空字符串;注意只是保存 文件的hash 地址；而不是 URL 地址
        example: ''''''
        type: string
      createdAt:
        description: 创建时间
        type: string
      id:
        description: 评论ID
        type: integer
      momentId:
        description: 朋友圈记录ID
        type: integer
      nickname:
        description: 用户中文名称，默认为 ‘’，可后期通过修改用户信息设置
        type: string
      status:
        description: 评论状态，0-正常，1-删除
        type: integer
      updatedAt:
        description: 更新时间
        type: string
      userId:
        description: 评论者用户ID
        type: integer
      visible:
        description: 评论内容
        type: string
    required:
    - avatar
    type: object
  model.MomentLikeResponse:
    properties:
      avatar:
        description: 用户头像、创建时；为空字符串;注意只是保存 文件的hash 地址；而不是 URL 地址
        example: ''''''
        type: string
      createdAt:
        description: 创建时间
        type: string
      id:
        description: 点赞 ID
        type: integer
      momentId:
        description: 朋友圈记录 ID
        type: integer
      nickname:
        description: 用户中文名称，默认为 ‘’，可后期通过修改用户信息设置
        type: string
      userId:
        description: 点赞者 ID
        type: integer
    required:
    - avatar
    type: object
  model.MomentPrivacy:
    properties:
      hideMine:
        description: 不让TA看我的朋友圈，0-不屏蔽，1-屏蔽
        type: boolean
      hideTheir:
        description: 我不看TA的朋友圈，0-不屏蔽，1-屏蔽
        type: boolean
      id:
        description: 隐私设置ID
        type: integer
      targetId:
        description: 目标用户ID
        type: integer
      userId:
        description: 用户ID
        type: integer
    type: object
  model.MomentVisbileStatus:
    enum:
    - 0
    - 1
    - 2
    - 3
    type: integer
    x-enum-varnames:
    - MomentVisibleStatusPublic
    - MomentVisibleStatusFriendsOnly
    - MomentVisibleStatusSelfOnly
    - MomentVisibleStatusPartialFriends
...
  model.UserResponse:
    properties:
      avatar:
        description: 用户头像、创建时；为空字符串;注意只是保存 文件的hash 地址；而不是 URL 地址
        example: ''''''
        type: string
      createAt:
        type: string
      id:
        description: 用户 ID;其类型是一个 uint 类型
        type: integer
      nickname:
        description: 用户中文名称，默认为 ‘’，可后期通过修改用户信息设置
        type: string
      token:
        description: 认证令牌
        type: string
      updateAt:
        type: string
      username:
        description: 用户账户；创建时使用；后期不可修改。
        example: whx
        type: string
    required:
    - avatar
    - id
    - token
    - username
    type: object
  utils.KResponse:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: |-
    一个 IM 服务器
    支持用户注册、登录、好友管理、群组聊天
    基于 Go + Gin + GORM + Redis
    目前访问服务器任何接口: localhost:8080/api 前端可用先使用代理设置
    使用文件 hash 获取文件地址可用先固定: localhost:8080/api/file/:hash 前端可用看作做;方便后期迁移
  title: go-net
  version: "1.0"
paths:
...
  /moments:
    get:
      parameters:
      - default: 20
        description: 限制条数
        in: query
        name: limit
        type: integer
      - default: 0
        description: 偏移量
        in: query
        name: offset
        type: integer
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/utils.KResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.UserResponse'
              type: object
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/utils.KResponse'
      summary: 获取朋友圈列表 (首页刷的流,当前登陆用户 + 好友)
      tags:
      - 朋友圈
    post:
      parameters:
      - description: 注册请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.Moment'
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/utils.KResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.Moment'
              type: object
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/utils.KResponse'
      summary: 发布朋友圈;仅取 Elements,Visbile 字段.
      tags:
      - 朋友圈
  /moments/:id:
    delete:
      parameters:
      - description: 朋友圈ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/utils.KResponse'
            - properties:
                data:
                  type: integer
              type: object
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/utils.KResponse'
      summary: 删除朋友圈
      tags:
      - 朋友圈
  /moments/:userID:
    get:
      parameters:
      - default: 20
        description: 限制条数
        in: query
        name: limit
        type: integer
      - default: 0
        description: 偏移量
        in: query
        name: offset
        type: integer
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/utils.KResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Moment'
                  type: array
              type: object
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/utils.KResponse'
      summary: 看某个人的朋友圈
      tags:
      - 朋友圈
  /moments/comments/:id:
    delete:
      parameters:
      - description: 评论ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/utils.KResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.MomentCommentsResponse'
              type: object
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/utils.KResponse'
      summary: 删除一条朋友圈评论
      tags:
      - 朋友圈/评论
      - 评论
    get:
      parameters:
      - description: 请求体
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/utils.KResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.MomentCommentsResponse'
                  type: array
              type: object
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/utils.KResponse'
      summary: 获取一条朋友的评论列表
      tags:
      - 朋友圈/评论
    post:
      parameters:
      - description: 评论体
        in: body
        name: reqeust
        required: true
        schema:
          $ref: '#/definitions/model.MomentComments'
      - description: 朋友圈ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/utils.KResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.MomentCommentsResponse'
              type: object
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/utils.KResponse'
      summary: 评论一条朋友圈;只取 MomentID,UserID,Content 字段
      tags:
      - 朋友圈/评论
    put:
      parameters:
      - description: 评论ID
        in: path
        name: id
        required: true
        type: integer
      - description: 评论体
        in: body
        name: reqeust
        required: true
        schema:
          $ref: '#/definitions/model.MomentComments'
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/utils.KResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.MomentCommentsResponse'
              type: object
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/utils.KResponse'
      summary: 修改一条评论列表;只取 Content 字段
      tags:
      - 朋友圈/评论
  /moments/likes/:id:
    delete:
      parameters:
      - description: 朋友圈ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/utils.KResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.Moment'
              type: object
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/utils.KResponse'
      summary: 取消点赞
      tags:
      - 朋友圈/点赞
    get:
      parameters:
      - description: 请求体
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/utils.KResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.MomentLikeResponse'
                  type: array
              type: object
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/utils.KResponse'
      summary: 获取一条朋友的点赞列表
      tags:
      - 朋友圈/点赞
    post:
      parameters:
      - description: 朋友圈ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/utils.KResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.Moment'
              type: object
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/utils.KResponse'
      summary: POST /api/moments/:id/like —— 点赞。传 id。
      tags:
      - 朋友圈/点赞
  /moments/privacy/:userID:
    get:
      parameters:
      - description: 好友的用户ID
        in: path
        name: userID
        required: true
        type: integer
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/utils.KResponse'
            - properties:
                data:
                  type: integer
              type: object
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/utils.KResponse'
      summary: 查和某个人的屏蔽设置。传 targetId。
      tags:
      - 朋友圈/隐私模块
    post:
      parameters:
      - description: 好友的用户ID
        in: path
        name: userID
        required: true
        type: integer
      - description: 请求体
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.MomentPrivacy'
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/utils.KResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.MomentPrivacy'
              type: object
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/utils.KResponse'
      summary: 设置屏蔽。传 hide_their（我不看TA的）、hide_mine（不让TA看我的）。
      tags:
      - 朋友圈/隐私模块
...
swagger: "2.0"